feat(comms): Executor config, fix start_listening
Allow configuration of a custom executor to use in libp2p::Swarm and
for spawning the `SwarmTask`.

Change `start_listening` to always require a Multiaddress, instead of
optionally using an OS-assigned one. This is necessary because a
transport may be used where `/ip4/0.0.0.0/tcp/0` is not be supported.

Adjust tests to the changes.