fix(p2p): fix EventChannel, add tests (#293)
Refactor Sink and Stream implementation for EventChannel.
The Sink implementation allows to write a message to the channel (blocking or non-blocking).
In case of an non-blocking Sink, the Stream implementation has to be continuously polled. The polling drives the sending of messages once the inner channel has the capacity to receive it (in the SwarmTask, this polling is done in the central loop in Stream::run).